Extract VideoXBlock mixins into mixins.py (#166)
* Move VideoXBlock mixins into mixins.py
* Remove few pylint-disable flags from mixins
* Reduce complexity of PlaybackStateMixin.player_state getter
* Add TranscriptsMixinTests tests
* Add VideoXBlockTestBase.test_mark_untinished_raises_assertion_error
